requirements.txt
chromadb
requests
beautifulsoup4
streamlit
python-dotenv
lxml
sentence-transformers
groq